fix prototype of strncasecmp
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@28154 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/cartconv.c
@@ -232,7 +232,7 @@ static const unsigned char charmap[] = {
     '\370', '\371', '\372', '\373', '\374', '\375', '\376', '\377',
 };
 
-int strncasecmp(const char *s1, const char *s2, int n)
+int strncasecmp(const char *s1, const char *s2, size_t n)
 {
     unsigned char u1, u2;
