+++ vice/src/util.c
@@ -793,7 +793,7 @@ int strcasecmp(const char *s1, const char *s2)
 #endif
 
 #if !defined HAVE_STRNCASECMP
-int strncasecmp(const char *s1, const char *s2, unsigned int n)
+int strncasecmp(const char *s1, const char *s2, size_t n)
 {
     int c1, c2;
